order/requirements.txt
fastapi[standard]
pymongo
pydantic[email]
motor
grpcio
protobuf